+++ src/GRA.Data/Model/DynamicAvatarBundle.cs
@@ -11,6 +11,8 @@ namespace GRA.Data.Model
         [MaxLength(255)]
         public string Name { get; set; }
         public bool CanBeUnlocked { get; set; }
+        public bool IsDeleted { get; set; }
+        public bool HasBeenAwarded { get; set; }
 
         public ICollection<DynamicAvatarBundleItem> DynamicAvatarBundleItems { get; set; }
     }