+++ src/GRA.Domain.Model/DynamicAvatarItem.cs
@@ -7,6 +7,7 @@ namespace GRA.Domain.Model
         [Required]
         public int DynamicAvatarLayerId { get; set; }
         public int DynamicAvatarLayerPosition { get; set; }
+        public string DynamicAvatarLayerName { get; set; }
 
         [MaxLength(255)]
         [Required]