+++ src/GRA.Domain.Repository/ITriggerRepository.cs
@@ -27,5 +27,7 @@ namespace GRA.Domain.Repository
         Task<bool> SecretCodeInUseAsync(int siteId, string secretCode);
         Task<Trigger> GetByBadgeIdAsync(int badgeId);
         Task RemoveUserTriggerAsync(int userId, int triggerId);
+        Task<ICollection<Trigger>> GetTriggersAwardingBundleAsync(int bundleId);
+        Task<bool> BundleIsInUseAsync(int bundleId);
     }
 }