+++ src/GRA.Domain.Service/ActivityService.cs
@@ -1102,7 +1102,7 @@ namespace GRA.Domain.Service
         private async Task AwardUserBundle(int userId, int bundleId,
             bool userIdIsCurrentUser = false)
         {
-            var bundle = await _dynamicAvatarBundleRepository.GetByIdAsync(bundleId);
+            var bundle = await _dynamicAvatarBundleRepository.GetByIdAsync(bundleId, false);
             if (bundle.DynamicAvatarItems.Count > 0)
             {
                 var loggingUser = (userIdIsCurrentUser ? userId : GetActiveUserId());
@@ -1138,6 +1138,12 @@ namespace GRA.Domain.Service
                     AvatarBundleId = bundleId,
                     Description = $"You unlocked the <strong>{bundle.Name}</strong> avatar bundle!"
                 });
+
+                if (!bundle.HasBeenAwarded)
+                {
+                    bundle.HasBeenAwarded = true;
+                    await _dynamicAvatarBundleRepository.UpdateSaveAsync(loggingUser, bundle);
+                }
             }
         }
     }