+++ src/GRA.Domain.Service/TriggerService.cs
@@ -11,14 +11,16 @@ namespace GRA.Domain.Service
     public class TriggerService : BaseUserService<TriggerService>
     {
         private readonly IBranchRepository _branchRepository;
+        private readonly IDynamicAvatarBundleRepository _dynamicAvatarBundleRepository;
         private readonly IEventRepository _eventRepository;
         private readonly IProgramRepository _programRepository;
         private readonly ISystemRepository _systemRepository;
         private readonly ITriggerRepository _triggerRepository;
-        public TriggerService(ILogger<TriggerService> logger, 
+        public TriggerService(ILogger<TriggerService> logger,
             GRA.Abstract.IDateTimeProvider dateTimeProvider,
             IUserContextProvider userContextProvider,
             IBranchRepository branchRepository,
+            IDynamicAvatarBundleRepository dynamicAvatarBundleRepository,
             IEventRepository eventRepository,
             IProgramRepository programRepository,
             ISystemRepository systemRepository,
@@ -26,6 +28,8 @@ namespace GRA.Domain.Service
         {
             SetManagementPermission(Permission.ManageTriggers);
             _branchRepository = Require.IsNotNull(branchRepository, nameof(branchRepository));
+            _dynamicAvatarBundleRepository = Require.IsNotNull(dynamicAvatarBundleRepository,
+                nameof(dynamicAvatarBundleRepository));
             _eventRepository = Require.IsNotNull(eventRepository, nameof(eventRepository));
             _programRepository = Require.IsNotNull(programRepository, nameof(programRepository));
             _systemRepository = Require.IsNotNull(systemRepository, nameof(systemRepository));
@@ -160,12 +164,12 @@ namespace GRA.Domain.Service
             return await _triggerRepository.GetTriggerDependentsAsync(triggerId);
         }
 
-        public async Task<bool> SecretCodeInUseAsync (string username)
+        public async Task<bool> SecretCodeInUseAsync(string username)
         {
             return await _triggerRepository.SecretCodeInUseAsync(GetCurrentSiteId(), username);
         }
 
-        public async Task<Trigger> GetByBadgeIdAsync (int badgeId)
+        public async Task<Trigger> GetByBadgeIdAsync(int badgeId)
         {
             return await _triggerRepository.GetByBadgeIdAsync(badgeId);
         }
@@ -221,6 +225,17 @@ namespace GRA.Domain.Service
             else if (!(trigger.Points > 0))
             {
                 throw new GraException("No criteria selected.");
+            }
+
+            if (trigger.AwardAvatarBundleId.HasValue)
+            {
+                var bundle = await _dynamicAvatarBundleRepository
+                    .GetByIdAsync(trigger.AwardAvatarBundleId.Value, false);
+
+                if (bundle == null || bundle.CanBeUnlocked == false)
+                {
+                    throw new GraException("Invalid Avatar Bundle selection.");
+                }
             }
         }
     }