+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Add achieved at property to Users
 - Add WebSocket-based reporting interface to avoid timeouts on report execution
 - Add Current Status, Participant Progress, and Registrations/Achievers reports
+- Add Mission Control interface for avatar bundles
 
 ### Changed
 - Change submit buttons to be disabled on click for challenge and household pages